// src/shared/api/@generated/model/inquiryDetailResponse.ts
/**
 * Generated by orval v8.11.0 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */
import type { InquiryAnswerResponse } from './inquiryAnswerResponse.ts';

export interface InquiryDetailResponse {
  id: number;
  title: string;
  type: string;
  content: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  /** 어드민 답변 (없으면 null) */
  answer?: InquiryAnswerResponse;
}